+++ libs/api/a-vehicles/src/lib/services/vehicles.service.ts
@@ -2,6 +2,7 @@ import { CreateVehicleDto, filterCommonQuery, GetVehiclesDto, UpdateVehicleDto, 
 import {
   createCollectionResponse,
   IvtCollectionResponse,
+  RevisionStatus,
   sortDirection,
   transformFolio,
   User,
@@ -85,8 +86,10 @@ export class VehiclesService {
   }
 
   async updateVehicle(updateVehcleDto: UpdateVehicleDto): Promise<Vehicle> {
-    const updatedVehicle = await this.vehicleRepository.save(updateVehcleDto);
-    return updatedVehicle;
+    const preloadedGuarantee = await this.vehicleRepository.preload(updateVehcleDto);
+    await preloadedGuarantee.save();
+    await preloadedGuarantee.reload();
+    return preloadedGuarantee;
   }
 
   async deleteVehicle(id: number): Promise<Vehicle> {
@@ -110,8 +113,8 @@ export class VehiclesService {
 
   @Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
   async updateVehicleStatusFromRevisions() {
-    const query = this.vehicleRepository.createQueryBuilder('vehicle');
-    await query
+    await this.vehicleRepository
+      .createQueryBuilder('vehicle')
       .update()
       .set({
         status: VehicleStatus.needsRevision,
@@ -120,15 +123,16 @@ export class VehiclesService {
         `NOT EXISTS(
           SELECT NULL
           FROM guarantees,
-               revisions
+               revisions,
+               vehicles
           WHERE (guarantees."vehicleId" = vehicles.id
               AND guarantees."endDate" > CURRENT_DATE - INTERVAL '0 days')
              OR (revisions."vehicleId" = vehicles.id
               AND revisions."createdAt" > CURRENT_DATE - INTERVAL '3 months'
-              AND revisions.status = '1')
-      )`
+              AND revisions.status = :revisionStatus )
+              AND vehicles.status != :vehicleStatus )`,
+        { vehicleStatus: VehicleStatus.notElegible, revisionStatus: RevisionStatus.elegible }
       )
-      .andWhere('vehicle.status != :status', { status: VehicleStatus.notElegible })
       .execute();
   }
 }