+++ libs/ui/u-guarantees/src/lib/components/guarantee-form/guarantee-form.component.ts
@@ -10,8 +10,17 @@ import {
 } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 import { ApsValidators } from '@arphase/ui';
-import { Guarantee, isVehicleElegible, IvtCollectionResponseInfo, PersonTypes, Select, Vehicle } from '@ivt/c-data';
+import {
+  Guarantee,
+  isVehicleElegible,
+  IvtCollectionResponseInfo,
+  PersonTypes,
+  Select,
+  UserRoles,
+  Vehicle,
+} from '@ivt/c-data';
 import { filterNil, RfcValidatorTypes } from '@ivt/c-utils';
+import { REQUIRED_ROLES } from '@ivt/u-state';
 import { createAddressForm, IvtFormComponent } from '@ivt/u-ui';
 import { createVehicleForm } from '@ivt/u-vehicles';
 import { QueryParams } from '@ngrx/data';
@@ -59,11 +68,11 @@ export function createGuaranteeForm(): FormGroup {
   templateUrl: './guarantee-form.component.html',
   styleUrls: ['./guarantee-form.component.less'],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  providers: [{ provide: REQUIRED_ROLES, useValue: [UserRoles.superAdmin] }],
 })
 export class GuaranteeFormComponent extends IvtFormComponent<Guarantee> implements OnChanges, AfterViewInit {
   @Input() productOptions: Select[] = [];
   @Input() companyOptions: Select[] = [];
-  @Input() showCompanyInput: boolean;
   @Input() vehicle: Vehicle;
   @Input() currentVehicle: Vehicle;
   @Input() error: string;
@@ -103,6 +112,10 @@ export class GuaranteeFormComponent extends IvtFormComponent<Guarantee> implemen
     return !this.currentVehicle || isVehicleElegible(this.currentVehicle);
   }
 
+  get disableSubmit(): boolean {
+    return !this.isElegible || !!this.error;
+  }
+
   ngOnChanges(changes: SimpleChanges) {
     if (changes.form && this.form) {
       this.client
@@ -115,10 +128,6 @@ export class GuaranteeFormComponent extends IvtFormComponent<Guarantee> implemen
 
     if (changes.isEditable && this.item) {
       this.isEditable ? this.form.enable() : this.form.disable();
-    }
-
-    if (changes.showCompanyInput) {
-      this.showCompanyInput ? this.form.get('companyId').enable() : this.form.get('companyId').disable();
     }
   }
 
@@ -139,12 +148,14 @@ export class GuaranteeFormComponent extends IvtFormComponent<Guarantee> implemen
     this.showPhysicalForm = value === PersonTypes[PersonTypes.physical];
     this.showMoralForm = value === PersonTypes[PersonTypes.moral];
 
-    if (this.showPhysicalForm) {
-      this.moralInfoForm.disable();
-      this.physicalInfoForm.enable();
-    } else {
-      this.moralInfoForm.enable();
-      this.physicalInfoForm.disable();
+    if (this.isEditable) {
+      if (this.showPhysicalForm) {
+        this.moralInfoForm.disable();
+        this.physicalInfoForm.enable();
+      } else {
+        this.moralInfoForm.enable();
+        this.physicalInfoForm.disable();
+      }
     }
   }