+++ libs/ui/u-guarantees/src/lib/containers/guarantee-form-container/guarantee-form-container.component.ts
@@ -1,12 +1,11 @@
 import { ChangeDetectionStrategy, Component, OnDestroy, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Company, Guarantee, UserRoles } from '@ivt/c-data';
-import { filterExisting, filterNil } from '@ivt/c-utils';
+import { filterNil } from '@ivt/c-utils';
 import {
   CompanyCollectionService,
   fromVehicles,
   getAuthUserCompanyIdState,
-  getAuthUserRoleState,
   getVehiclesErrorState,
   getVehiclesVehicleState,
   GuaranteeCollectionService,
@@ -22,7 +21,8 @@ import { QueryParams } from '@ngrx/data';
 import { select, Store } from '@ngrx/store';
 import { omit } from 'lodash-es';
 import { NzMessageService } from 'ng-zorro-antd/message';
-import { map, take, takeUntil } from 'rxjs/operators';
+import { combineLatest } from 'rxjs';
+import { map, takeUntil } from 'rxjs/operators';
 
 import { createGuaranteeForm } from '../../components/guarantee-form/guarantee-form.component';
 
@@ -37,14 +37,19 @@ export class GuaranteeFormContainerComponent extends IvtFormContainerComponent<G
   successUrl = '/spa/guarantees';
   createSuccessMessage = 'La garantía se ha creado con éxito';
   updateSuccessMessage = 'La garantía se ha actualizado con éxito';
-  isEditable$ = this.permissionService.hasUpdatePermission([UserRoles.superAdmin, UserRoles.agencyUser]);
+  isEditable$ = combineLatest([
+    this.permissionService.hasCreatePermission([UserRoles.superAdmin, UserRoles.agencyUser]),
+    this.permissionService.hasUpdatePermission([UserRoles.superAdmin]),
+    this.route.url,
+  ]).pipe(
+    map(([create, update, url]) => {
+      const createRoute = url.find(segment => segment.path === 'new');
+      return createRoute ? create : update;
+    })
+  );
   productOptions$ = this.productCollectionService.options$;
   companyId$ = this.store.pipe(select(getAuthUserCompanyIdState));
   vehicle$ = this.vehicleCollectionService.currentItem$;
-  showCompanyInput$ = this.store.pipe(
-    select(getAuthUserRoleState),
-    map(role => role === UserRoles[UserRoles.superAdmin])
-  );
   companyOptions$ = this.companyCollectionService.options$;
   companiesInfo$ = this.companyCollectionService.store.pipe(
     select(this.companyCollectionService.selectors.selectCollection),
@@ -61,16 +66,13 @@ export class GuaranteeFormContainerComponent extends IvtFormContainerComponent<G
     private permissionService: PermissionService,
     private store: Store<IvtState>,
     private companyCollectionService: CompanyCollectionService,
-    private vehicleCollectionService: VehicleCollectionService
+    private vehicleCollectionService: VehicleCollectionService,
+    private route: ActivatedRoute
   ) {
     super(guaranteeCollectionService, router, messageService);
   }
 
   ngOnInit() {
-    this.showCompanyInput$.pipe(take(1), filterExisting()).subscribe(() => {
-      this.companyCollectionService.getWithQuery({});
-    });
-
     this.store
       .pipe(select(selectQueryParam('vehicleId')), takeUntil(this.destroy$), filterNil())
       .subscribe(id => this.vehicleCollectionService.getByKey(Number(id)));