+++ libs/ui/u-state/src/lib/permissions/directives/no-create-permission.directive.spec.ts
@@ -1,0 +1,8 @@
+import { NoCreatePermissionDirective } from './no-create-permission.directive';
+
+describe('NoCreatePermissionDirective', () => {
+  it('should create an instance', () => {
+    const directive = new NoCreatePermissionDirective();
+    expect(directive).toBeTruthy();
+  });
+});