+++ libs/ui/u-state/src/lib/permissions/directives/no-delete-permission.directive.spec.ts
@@ -1,0 +1,8 @@
+import { NoDeletePermissionDirective } from './no-delete-permission.directive';
+
+describe('NoDeletePermissionDirective', () => {
+  it('should create an instance', () => {
+    const directive = new NoDeletePermissionDirective();
+    expect(directive).toBeTruthy();
+  });
+});